WIP: update apiURL of clients for internal routing in production
This commit updates the apiURL of clients to utilize internal routing in the production environment.
By upgrading to the cli version 0.0.9 and server version 0.0.14, this enhancement becomes possible.

The changes made in this commit will ensure that clients access the necessary APIs internally,
optimizing the routing process and improving overall performance.

Note: Upgrading to cli version 0.0.9 and server version 0.0.14 is required
for this update to work correctly.